requirements-tastypie.txt
python-mimeparse
python-dateutil
lxml
defusedxml
pyyaml
django-tastypie
